requirements/requirements.txt
Django==1.11.3
django-annoying==0.10.3
django-autocomplete-light==3.2.9
django-bootstrap-breadcrumbs==0.8.2
django-crispy-forms==1.6.1
django-filter==1.0.4
django-tables2==1.10.0
lxml==3.8.0
MarkupSafe==1.0
psycopg2==2.7.1
pytz==2017.2
six==1.10.0
